Add rubin-alert-sim command
Add a new setup.py file which makes this project pip installable. When
installing, install a rubin-alert-sim executable. This is a command
line application with only one subcommand: rubin-alert-sim
publish-file will send a file to a Kafka broker.

This CLI is eventually going to house more subcommands. For now, just
set up the scaffolding.