Test the setup.py script in Github Actions CI
Add a github workflow which makes sure that this project is pip
installable on Python 3.6, 3.7, and 3.8. In addition, check that the
resulting rubin-alert-sim executable can publish a file without
throwing an error.

Why not Python 3.5?
===================

It turns out that version 1.9.2.1 of avro-python3 is not installable
with the Python 3.5 Github Actions runner if it is mentioned as a
dependency in a setup.py file. It appears to create an egg file with
an unknown name, and dies. I haven't been able to reproduce this
locally, but it happens on every try inside github. I have a minimal
reproducer at https://github.com/spenczar/demo-avro-install-bug/.

The bug doesn't seem to happen if one explicitly installs avro-python3
with pip, only when installing as a dependency. I don't understand
this and won't pursue it further.