Add create-stream subcommand
Replace the publish-file command with a create-stream command. This
publishes data from an alert file to Kafka, but it also sets header
values indicating a time offset for each alert message.

In a future commit, this time offset can be consumed to replay the
stream at its original pace without needing to deserialize any
messages.

Along the way, add a Kafka client wrapper which manages topic creation
and deletion, subscription, and administration. Also write a bunch of
tests of this.