Use wget instead of curl to download datasets in Makefile
wget will use the HTTP headers describing the last-modified time of a file. It
sets the modification time of the download file accordingly. This results in a
consistent set of files for the data/ directory, which in turn ensures that
docker layer caching works, reducing build time and docker image size.